requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
backports.zoneinfo==0.2.1
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
chromedriver-binary==109.0.5414.25.0
cnum==0.0.6
Django==4.1.3
django-environ==0.9.0
exceptiongroup==1.0.4
google-api-core==2.11.0
google-api-python-client==2.70.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.57.0
h11==0.14.0
httplib2==0.21.0
idna==3.4
isodate==0.6.1
newsapi-python==0.2.6
oauthlib==3.2.2
outcome==1.2.0
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
trio==0.22.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.13
wsproto==1.2.0
